feat: homepage SEO optimization
